backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
sqlalchemy
google-generativeai  # placeholder - install actual official package if named differently
pydantic-settings